Add files via upload
Run recordForRegistration.py file and click on the register button after that it will star recordinig, and next time when you run recordForRegistration.py agin and click on the "pridict voice" button then click on " record voice" button for recording voice for prediction then click on "Predict" button. Model is not trained well cause it's a demo one next time going to use transfer learning on top of this project. for better accuracy.